# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |
    This OpenAPI document describes the WASAText API.
  version: "0.1.0"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://127.0.1:8080/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Use the user ID as the bearer token.
        Example: `Authorization: 12345`
      bearerFormat: userId
  responses:
    BadRequest:
      description: Bad Request - The request was invalid or malformed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidData:
              value:
                code: 400
                message: "Invalid request data"
    Unauthorized:
      description: Unauthorized - Authentication is required or has failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            unauthorized:
              value:
                code: 401
                message: "Authentication required"
    Forbidden:
      description: Forbidden - The request is understood but access is denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              value:
                code: 403
                message: "Access denied"
    NotFound:
      description: Not Found - The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            notFound:
              value:
                code: 404
                message: "Resource not found"
    InternalServerError:
      description: Internal Server Error - An unexpected error occurred
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            serverError:
              value:
                code: 500
                message: "Internal server error"
  schemas:
    ErrorResponse:
      type: object
      description: Error response
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          description: Error code, usually an HTTP status code
          example: 404
        message:
          type: string
          description: Error message
          example: "User not found"
    ImageReference:
      type: object
      description: Reference to an image file
      required:
      - uuid
      properties:
        uuid:
          type: string
          format: uuid
          description: The UUID reference to the image file
          example: "550e8400-e29b-41d4-a716-446655440000"
          minLength: 36
          maxLength: 36
          pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
    PublicUser:
      type: object
      description: A user object representing a public user profile
      required:
        - username
      properties:
        username:
          type: string
          description: The user’s display name
          example: Maria
          minLength: 3
          maxLength: 16
          pattern: '^[a-zA-Z0-9_]+$'
        photo:
          $ref: '#/components/schemas/ImageReference'
    PrivateUser:
      allOf:
        - $ref: '#/components/schemas/PublicUser'
        - type: object
          description: An internal user object
          required:
            - id
          properties:
            id:
              type: integer
              format: int64
              description: Database‐generated user ID
              example: 1
              maximum: 9223372036854775807
              minimum: 1
...
paths:
  /session:
    post:
      tags:
        - login
      summary: Log in or create a user
      description: If the user does not exist, it will be created.
      operationId: doLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicUser'
      responses:
        '201':
          description: User log-in action successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateUser'
              examples:
                success:
                  value:
                    id: 1
                    username: "Maria"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /upload:
    post:
      tags:
        - login
      summary: Upload an image
      description: Uploads an image and returns its UUID reference.
      operationId: uploadImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              description: Request body containing the image file
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
                  description: The image file to upload
                  minLength: 1
                  maxLength:  10485760 # 10 MB
              required:
                - imageFile
      responses:
        '201':
          description: Image uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageReference'
              examples:
                success:
                  value:
                    uuid: "550e8400-e29b-41d4-a716-446655440000"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me/username:
    put:
      tags:
        - user
      summary: Change your username
      description: Sets the user’s display name.
      operationId: setMyUserName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Request body containing the new username
              type: object
              properties:
                name:
                  type: string
                  description: New display name
                  example: Maria
                  minLength: 3
                  maxLength: 16
                  pattern: '^[a-zA-Z0-9_]+$'
              required:
                - name
      responses:
        '200':
          description: Username updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateUser'
              examples:
                success:
                  value:
                    id: 1
                    username: "Maria"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /me/photo:
    put:
      tags:
        - user
      summary: Change your photo
      description: Sets the user’s profile photo.
      operationId: setMyPhoto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body containing the image UUID reference
              properties:
                image:
                  $ref: '#/components/schemas/ImageReference'
              required:
                - image
      responses:
        '200':
          description: Photo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PrivateUser'
              examples:
                success:
                  value:
                    id: 1
                    username: "Maria"
                    photo:
                      uuid: "550e8400-e29b-41d4-a716-446655440000"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /conversations/{conversationId}/name:
    parameters:
      - name: conversationId
        description: Conversation identifier to rename the group
        in: path
        required: true
        schema:
          type: integer
    put:
      tags:
        - group
      summary: Rename a group
      description: Sets the group conversation’s name.
      operationId: setGroupName
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body containing the new group name
              properties:
                name:
                  type: string
                  description: New group name
                  example: Group 1
                  minLength: 1
                  maxLength: 16
                  pattern: '^[a-zA-Z0-9_ ]+$'
              required:
                - name
      responses:
        '200':
          description: Group renamed
          content:
            application/json:
              schema:
                type: string
                description: The new name of the group
              examples:
                success:
                  value: "Group 1"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /conversations/{conversationId}/photo:
    parameters:
      - name: conversationId
        description: Conversation identifier to change the photo
        in: path
        required: true
        schema:
          type: integer
    put:
      tags:
        - group
      summary: Change group photo
      description: Sets the group conversation’s display photo.
      operationId: setGroupPhoto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body containing the image UUID reference
              properties:
                image:
                  $ref: '#/components/schemas/ImageReference'
              required:
                - image
      responses:
        '200':
          description: Group photo updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageReference'
              examples:
                success:
                  value:
                    uuid: "550e8400-e29b-41d4-a716-446655440000"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /conversations/{conversationId}/participants:
    parameters:
      - name: conversationId
        description: Conversation identifier to manage participants
        in: path
        required: true
        schema:
          type: integer
    post:
      tags:
        - group
      summary: Add participants
      description: Adds users to a group conversation.
      operationId: addToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Request body containing the list of users to add
              type: object
              properties:
                participants:
                  type: array
                  description: Users to add
                  items:
                    $ref: '#/components/schemas/PublicUser'
                  minItems: 1
                  maxItems: 1000
              required:
                - participants
      responses:
        '204':
          description: Participants added successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - group
      summary: Leave group
      description: Current user leaves the group conversation.
      operationId: leaveGroup
      responses:
        '204':
          description: Left group successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...